+++ src/test/java/com/mageddo/dnsproxyserver/solver/remote/circuitbreaker/application/CircuitExecutionsAsHealthCheckerTest.java
@@ -1,5 +1,6 @@
 package com.mageddo.dnsproxyserver.solver.remote.circuitbreaker.application;
 
+import com.mageddo.dnsproxyserver.solver.remote.circuitbreaker.canaryratethreshold.CircuitExecutionsAsHealthChecker;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
@@ -8,6 +9,7 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import testing.templates.solver.remote.ResultSupplierTemplates;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 @ExtendWith(MockitoExtension.class)
 class CircuitExecutionsAsHealthCheckerTest {
@@ -27,5 +29,11 @@ class CircuitExecutionsAsHealthCheckerTest {
     this.obj.isHealthy();
 
     assertEquals(1, sup.getCalls());
+  }
+
+  @Test
+  void mustAnswerHealthWhenExecuteNeverCalled(){
+    final var healthy = this.obj.isHealthy();
+    assertTrue(healthy);
   }
 }
